fix: fix a critical winding bug
"Winding" refers to the specific way the points of a polyhedron are stored and referred to. It is important
because calculating some elements (such as surface normals) requires the winding of points to be either fully
clockwise or counterclockwise. A bug in the mesh generation function was causing winding to be incorrect.
That has been fixed.