feat: add mesh class
Implemented a simple mesh object that stores untextured mesh data as a group of vertices, indices, and normals.

The sample render of the lit sphere in main.cxx is updated to make use of this. The sphere generation algorithm
has also been moved as a debug function that is included along with the mesh class.